"+ new interface" is now an option
selecting the "+new interface" option opens urxvt running wifi-menu with
sudo. There doesn't seem to be a way to reasonably get the users
terminal binary name. so thats not going to work. For now it just
defaults to urxvt. for xterm or whatever other terminal, simply put a
symlink into some directory in your path